goit-ds-hw-03/pyproject.toml
[tool.poetry]
name = "goit-ds-hw-03"
version = "0.1.0"
description = "GoIT, Homework #3"
authors = ["artur.trotskyi <artur.trotskyi@ukr.net>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "mongo_homework" }]

[tool.poetry.dependencies]
python = "3.12.8"
pymongo = "^10.10.10.10"
python-dotenv = "^1.0.1"
faker = "^35.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
